feat: extract vedo plotting utilities
